Rearranging GUI for usability.
